Create Technical Page HTML
This is the HTML code snippet for a technical page, in this case, Python. All of the materials included are taken from other sites that I have left as links in the reference section at the end of the snippet. All of the code parts are written by me and as such, you can use it however you like.